+++ vhdllab-server/src/main/java/hr/fer/zemris/vhdllab/service/EntityManager.java
@@ -38,7 +38,7 @@ public interface EntityManager<T> {
 	 * @throws NullPointerException
 	 *             if <code>id</code> is <code>null</code>
 	 */
-	T load(Long id) throws ServiceException;
+	T load(Integer id) throws ServiceException;
 
 	/**
 	 * Returns <code>true</code> if an entity with specified identifier exists
@@ -53,7 +53,7 @@ public interface EntityManager<T> {
 	 * @throws NullPointerException
 	 *             if <code>id</code> is <code>null</code>
 	 */
-	boolean exists(Long id) throws ServiceException;
+	boolean exists(Integer id) throws ServiceException;
 
 	/**
 	 * Deletes an entity. If entity doesn't exist then this method will throw
@@ -66,6 +66,6 @@ public interface EntityManager<T> {
 	 * @throws NullPointerException
 	 *             if <code>id</code> is <code>null</code>
 	 */
-	void delete(Long id) throws ServiceException;
+	void delete(Integer id) throws ServiceException;
 
 }
