+++ vhdllab-server/src/main/java/hr/fer/zemris/vhdllab/servlets/methods/DoMethodDeleteProject.java
@@ -24,7 +24,7 @@ public class DoMethodDeleteProject extends AbstractRegisteredMethod {
      */
     @Override
     public void run(Method<Serializable> method, HttpServletRequest request) {
-		Long projectId = method.getParameter(Long.class, PROP_ID);
+        Integer projectId = method.getParameter(Integer.class, PROP_ID);
 		if (projectId == null) {
 			return;
 		}
