+++ vhdllab-server/src/main/java/hr/fer/zemris/vhdllab/servlets/methods/DoMethodExistsFile.java
@@ -24,7 +24,7 @@ public class DoMethodExistsFile extends AbstractRegisteredMethod {
      */
     @Override
     public void run(Method<Serializable> method, HttpServletRequest request) {
-		Long fileId = method.getParameter(Long.class, PROP_ID);
+        Integer fileId = method.getParameter(Integer.class, PROP_ID);
 		if (fileId == null) {
 			return;
 		}
