+++ vhdllab-server/src/main/java/hr/fer/zemris/vhdllab/servlets/methods/DoMethodExtractHierarchy.java
@@ -26,7 +26,7 @@ public class DoMethodExtractHierarchy extends AbstractRegisteredMethod {
      */
     @Override
     public void run(Method<Serializable> method, HttpServletRequest request) {
-		Long projectId = method.getParameter(Long.class, PROP_ID);
+        Integer projectId = method.getParameter(Integer.class, PROP_ID);
 		if (projectId == null) {
 			return;
 		}
